src/requirements/local.txt
# Extend base.txt
-r base.txt

# Add local extra requirements here (django-debug etc)
django-debug-toolbar
isort
pipdeptree
